@@ -1,4 +1,4 @@
-class CreateSpreeDeliverySlots < ActiveRecord::Migration
+class CreateSpreeDeliverySlots < SpreeExtension::Migration[4.2]
   def change
     create_table :spree_delivery_slots do |t|
       t.references :shipping_method, index: true
