@@ -1,4 +1,4 @@
-class CreateSpreeShipmentDeliverySlot < ActiveRecord::Migration
+class CreateSpreeShipmentDeliverySlot < SpreeExtension::Migration[4.2]
   def change
     create_table :spree_shipments_delivery_slots do |t|
       t.references :shipment, index: true
